TG-79: moved getProjects action to thunks.
