+++ src/redux/projects/thunks.js
@@ -1,0 +1,17 @@
+import { getProjectsPending, getProjectsSuccess, getProjectsError } from './actions';
+
+const getProjects = () => {
+  return (dispatch) => {
+    dispatch(getProjectsPending());
+    fetch(`${process.env.REACT_APP_API_URL}/projects`)
+      .then((response) => response.json())
+      .then((response) => {
+        dispatch(getProjectsSuccess(response.data));
+      })
+      .catch((error) => {
+        dispatch(getProjectsError(error.toString()));
+      });
+  };
+};
+
+export default getProjects;
